Phase 3 Task 17 cycle 3 hardening: GridSizing extraction + zero-cell content + chrome-inclusive intrinsic + narrowed scope
Closes 6 PR-#94 review findings (2 P1, 2 P2, 2 P3):

F1+F6 (P1): Extract src/NetPdf.Layout/Layouters/GridSizing.cs as the
single shared sizing service. Both BlockLayouter.PreMeasureGridRowExtent
and GridLayouter emission now call GridSizing.Resolve(...) — pre-measure
runs as a dry-run (emit: null) so diagnostics fire only at real emission.
Pre-F1 the pre-measure was length-only; auto-height wrappers collapsed
when rows used intrinsic / fr sizing and following siblings overlapped.
GridLayouter slimmed from ~1522 → ~512 LOC; PlacedItem/PlacementSpec/
PlacementKind moved with the placement code.

F2 (P1): Zero-sized cell content now dispatches with fallback extent
(inlineSize > 0 ? inlineSize : _contentInlineSize) instead of being
skipped. The outer item still emits at zero geometry; children layout
against the grid's content extent + visually overflow. The existing
LayoutGridZeroSizedCellContentSkipped001 diagnostic still surfaces the
condition.

F3 (P2): GridSizing.ItemOuterContribution now adds border + padding +
margin to the declared dimension before contributing to the intrinsic
track sum. An item with width: 100px; padding: 10px; border: 2px
contributes 124px to its column track (was 100).

F4 (P2): Narrowed cycle-3 scope claim. GridLayouter class XML doc
explicitly lists the 3 track kinds shipped (Length / Fr with §11.7.1
spec-correct floor / intrinsic via L19 approximation) AND the 5
deferrals NOT in cycle 3 (true intrinsic content measurement = L19,
§11.6 Maximize step = cycle 4, auto-track stretch = separate task,
box-sizing: border-box, indefinite-axis percentage resolution). New
docs/deferrals.md entry grid-track-sizing-cycle3-narrowed-scope
documents each deferral with trigger + removal condition.

F5 (P3): Reworded LayoutGridFrUnderIndefiniteApproximated001
diagnostic message — no longer says "cycle 3 will ship intrinsic
resolution" since this IS cycle 3. Now says it remains blocked on
full intrinsic / max-content content-size measurement (L19).

Tests: 5038 unit + 97 RealDocs + 4 AotJitParity green. No new test
additions — pure hardening of existing surfaces. AOT/JIT byte-parity
preserved at sha256 2942DD1E…30C3DE7.
